fix: atualiza estrutura do build fast api, alterando de api para routers e de cruds para use_cases.
