+++ {{cookiecutter.project_slug}}/core/management/commands/fastapi_managers/__init__.py
@@ -1,12 +1,12 @@
-from .build_api import ApiBuild
-from .build_cruds import CrudsBuild
+from .build_routers import RoutersBuild
+from .build_use_cases import UseCasesBuild
 from .build_docker import DockerBuild
 from .build_models import ModelsBuild
 from .build_schemas import SchemasBuild
 
 __all__ = [
-    "ApiBuild",
-    "CrudsBuild",
+    "RoutersBuild",
+    "UseCasesBuild",
     "DockerBuild",
     "ModelsBuild",
     "SchemasBuild",
