+++ {{cookiecutter.project_slug}}/core/management/commands/snippets/fastapi/model.txt
@@ -1,6 +1,6 @@
 import datetime
-import sqlalchemy as sa
-from sqlalchemy.sql.sqltypes import String, Integer, Boolean, Date, DateTime, Time
+from sqlalchemy import ForeignKey, String, Table, Column
+from sqlalchemy.sql.sqltypes import String, Integer, Boolean, Date, DateTime, Time, List, Float
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from core.database import Base
