+++ {{cookiecutter.project_slug}}/core/management/commands/snippets/fastapi_project/docker-dev.yml
@@ -14,12 +14,12 @@ services:
       - APP_NAME=TROCAR_PELO_NOME_DO_SISTEMA
       - APP_URL=http://localhost:8005
       - DB_DATABASE=TROCAR_PELO_NOME_DO_SISTEMA_db
-      - DB_PASSWORD=senha_padrao_deve_ser_mudada
+      - DB_PASSWORD=2LiyBoLHeHo5yGfxan8euHGIzEEzIs
       - DB_USERNAME=TROCAR_PELO_NOME_DO_SISTEMA_dbmanager_2LiyBoLHeHo5yG
       - DB_HOST=TROCAR_PELO_NOME_DO_SISTEMA_database
       - DB_PORT=5432
       - DB_CONNECTION=postgresql
-      - APP_SECRET=senha_padrao_deve_ser_mudada
+      - APP_SECRET=NYPmGeU11uQ96YQtX9UjOZ_EuAGsuJltGJt5XBrXt0HQhgShCP7KIWlLFgxtZDhl0z4
     build:
       context: .
       dockerfile: ./DockerfileDev
