+++ {{cookiecutter.project_slug}}/core/management/commands/snippets/fastapi_project/requirements.txt
@@ -1,11 +1,11 @@
 #
-# This file is autogenerated by pip-compile with Python 3.11
+# This file is autogenerated by pip-compile with Python 3.9
 # by the following command:
 #
-#    pip-compile '.\requirements.in'
+#    pip-compile requirements.in
 #
 alembic==1.11.1
-    # via -r .\requirements.in
+    # via -r requirements.in
 anyio==3.6.2
     # via starlette
 certifi==2023.5.7
@@ -14,16 +14,14 @@ charset-normalizer==3.1.0
     # via requests
 click==8.1.3
     # via uvicorn
-colorama==0.4.6
-    # via click
 dnspython==2.3.0
     # via email-validator
 ecdsa==0.18.0
     # via python-jose
 email-validator==2.0.0.post2
-    # via -r .\requirements.in
-fastapi==0.95.2
-    # via -r .\requirements.in
+    # via -r requirements.in
+fastapi==0.100.0
+    # via -r requirements.in
 greenlet==2.0.2
     # via sqlalchemy
 h11==0.14.0
@@ -38,9 +36,9 @@ mako==1.2.4
 markupsafe==2.1.2
     # via mako
 passlib==1.7.4
-    # via -r .\requirements.in
+    # via -r requirements.in
 psycopg2-binary==2.9.6
-    # via -r .\requirements.in
+    # via -r requirements.in
 pyasn1==0.5.0
     # via
     #   python-jose
@@ -48,13 +46,13 @@ pyasn1==0.5.0
 pydantic==1.10.7
     # via fastapi
 python-dotenv==1.0.0
-    # via -r .\requirements.in
+    # via -r requirements.in
 python-jose==3.3.0
-    # via -r .\requirements.in
+    # via -r requirements.in
 python-multipart==0.0.6
-    # via -r .\requirements.in
+    # via -r requirements.in
 requests==2.31.0
-    # via -r .\requirements.in
+    # via -r requirements.in
 rsa==4.9
     # via python-jose
 six==1.16.0
@@ -63,16 +61,18 @@ sniffio==1.3.0
     # via anyio
 sqlalchemy==2.0.15
     # via
-    #   -r .\requirements.in
+    #   -r requirements.in
     #   alembic
 starlette==0.27.0
     # via fastapi
 typing-extensions==4.5.0
     # via
     #   alembic
+    #   fastapi
     #   pydantic
     #   sqlalchemy
+    #   starlette
 urllib3==2.0.2
     # via requests
 uvicorn==0.22.0
-    # via -r .\requirements.in
+    # via -r requirements.in
