--- {{cookiecutter.project_slug}}/core/management/commands/snippets/fastapi_project/usuario/cruds.py
@@ -1,11 +1,0 @@
-from core.cruds import CRUDBase
-
-from .models import Usuario
-from .schemas import UsuarioCreate, UsuarioUpdate
-
-
-class CRUDUsuario(CRUDBase[Usuario, UsuarioCreate, UsuarioUpdate]):
-    pass
-
-
-usuario = CRUDUsuario(Usuario)
